layout refactor, delete unused imports
