+++ src/main/java/pl/bykowsi/kurs/tydzien1pd/controller/impl/ShopControllerPlus.java
@@ -20,7 +20,7 @@ public class ShopControllerPlus implements ShopController {
 
     @Override
     @EventListener(ApplicationReadyEvent.class)
-    public void start(){
+    public void start() {
         service.calculateFinalPrice();
     }
