+++ src/main/java/pl/bykowsi/kurs/tydzien1pd/service/impl/PlusShopService.java
@@ -5,9 +5,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
-import pl.bykowsi.kurs.tydzien1pd.model.Basket;
 import pl.bykowsi.kurs.tydzien1pd.model.Product;
 import pl.bykowsi.kurs.tydzien1pd.screeninfo.PrintInfo;
+import pl.bykowsi.kurs.tydzien1pd.service.Basket;
 import pl.bykowsi.kurs.tydzien1pd.service.ShopService;
 
 import java.math.BigDecimal;
@@ -26,7 +26,7 @@ public class PlusShopService implements ShopService {
     private static final BigDecimal hundred = BigDecimal.valueOf(100);
 
     @Autowired
-    public PlusShopService(@Value("${price.VAT}") Integer VAT, Basket basket,  MessageSource messageSource, @Value("${language.languageVersion}") String languageVersion) {
+    public PlusShopService(@Value("${price.VAT}") Integer VAT, Basket basket, MessageSource messageSource, @Value("${language.languageVersion}") String languageVersion) {
         this.VAT = BigDecimal.valueOf(VAT);
         this.basket = basket;
         this.sum = BigDecimal.ZERO;
@@ -36,15 +36,15 @@ public class PlusShopService implements ShopService {
 
 
     @Override
-    public void calculateFinalPrice(){
+    public void calculateFinalPrice() {
         List<Product> list = basket.getBasket();
 
-        list.forEach(p -> System.out.println(messageSource.getMessage("singleProductPosition", new Object[]{p.getName(),p.getPrice()}, Locale.forLanguageTag(languageVersion))));
+        list.forEach(p -> System.out.println(messageSource.getMessage("singleProductPosition", new Object[]{p.getName(), p.getPrice()}, Locale.forLanguageTag(languageVersion))));
         sum = list.stream()
                 .map(Product::getPrice)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
 
         BigDecimal grossSum = sum.multiply(hundred.add(VAT).divide(hundred));
-        PrintInfo.PlusPrintData(messageSource,sum,languageVersion,grossSum,VAT);
+        PrintInfo.PlusPrintData(messageSource, sum, languageVersion, grossSum, VAT);
     }
 }
