+++ src/main/java/pl/bykowsi/kurs/tydzien1pd/service/impl/ProShopService.java
@@ -5,9 +5,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
-import pl.bykowsi.kurs.tydzien1pd.model.Basket;
 import pl.bykowsi.kurs.tydzien1pd.model.Product;
 import pl.bykowsi.kurs.tydzien1pd.screeninfo.PrintInfo;
+import pl.bykowsi.kurs.tydzien1pd.service.Basket;
 import pl.bykowsi.kurs.tydzien1pd.service.ShopService;
 
 import java.math.BigDecimal;
@@ -27,7 +27,7 @@ public class ProShopService implements ShopService {
     private static final BigDecimal hundred = BigDecimal.valueOf(100);
 
     @Autowired
-    public ProShopService(@Value("${price.VAT}") Integer VAT,@Value("${price.discount}") Integer discount, Basket basket,   MessageSource messageSource, @Value("${language.languageVersion}") String languageVersion) {
+    public ProShopService(@Value("${price.VAT}") Integer VAT, @Value("${price.discount}") Integer discount, Basket basket, MessageSource messageSource, @Value("${language.languageVersion}") String languageVersion) {
         this.VAT = BigDecimal.valueOf(VAT);
         this.discount = BigDecimal.valueOf(discount);
         this.basket = basket;
@@ -37,9 +37,9 @@ public class ProShopService implements ShopService {
     }
 
     @Override
-    public void calculateFinalPrice(){
+    public void calculateFinalPrice() {
         List<Product> list = basket.getBasket();
-        list.forEach(p -> System.out.println(messageSource.getMessage("singleProductPosition", new Object[]{p.getName(),p.getPrice()}, Locale.forLanguageTag(languageVersion))));
+        list.forEach(p -> System.out.println(messageSource.getMessage("singleProductPosition", new Object[]{p.getName(), p.getPrice()}, Locale.forLanguageTag(languageVersion))));
         sum = list.stream()
                 .map(Product::getPrice)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
@@ -47,7 +47,7 @@ public class ProShopService implements ShopService {
         BigDecimal grossPrice = sum.multiply(hundred.add(VAT)).divide(hundred);
         BigDecimal discountedGrossPrice = grossPrice.multiply(hundred.subtract(discount)).divide(hundred);
 
-           PrintInfo.ProPrintData(messageSource,sum,languageVersion,grossPrice,VAT,discountedGrossPrice,100-discount.intValue());
+        PrintInfo.ProPrintData(messageSource, sum, languageVersion, grossPrice, VAT, discountedGrossPrice, 100 - discount.intValue());
     }
 
 }
