+++ src/main/java/pl/bykowsi/kurs/tydzien1pd/service/impl/StartShopService.java
@@ -5,9 +5,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
-import pl.bykowsi.kurs.tydzien1pd.model.Basket;
 import pl.bykowsi.kurs.tydzien1pd.model.Product;
 import pl.bykowsi.kurs.tydzien1pd.screeninfo.PrintInfo;
+import pl.bykowsi.kurs.tydzien1pd.service.Basket;
 import pl.bykowsi.kurs.tydzien1pd.service.ShopService;
 
 import java.math.BigDecimal;
@@ -35,11 +35,11 @@ public class StartShopService implements ShopService {
     @Override
     public void calculateFinalPrice() {
         List<Product> list = basket.getBasket();
-        list.forEach(p -> System.out.println(messageSource.getMessage("singleProductPosition", new Object[]{p.getName(),p.getPrice()}, Locale.forLanguageTag(languageVersion))));
+        list.forEach(p -> System.out.println(messageSource.getMessage("singleProductPosition", new Object[]{p.getName(), p.getPrice()}, Locale.forLanguageTag(languageVersion))));
         sum = list.stream()
-                    .map(Product::getPrice)
-                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                .map(Product::getPrice)
+                .reduce(BigDecimal.ZERO, BigDecimal::add);
 
-        PrintInfo.StartPrintData(messageSource,sum,languageVersion);
+        PrintInfo.StartPrintData(messageSource, sum, languageVersion);
     }
 }
