+++ src/main/resources/application-plus.properties
@@ -1,1 +1,2 @@
+#price data
 price.VAT = 26
