+++ src/main/resources/application-pro.properties
@@ -1,2 +1,3 @@
+#price data
 price.discount = 5
 price.VAT = 26
