+++ src/main/resources/application.properties
@@ -1,8 +1,11 @@
+#profile
 spring.profiles.active=start
 
+#price data
 price.MinPrice = 50
 price.MaxPrice = 301
 
-language.languageVersion = eng
+#language data
+language.languageVersion=eng
